Separated header and implementation
